Relacionamento Agente x Empresa

+++ src/main/java/com/squad03/flap/DTO/BuscaTarefa.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public record BuscaTarefa(
         Long id,
-        int empresaId,
+        Long empresaId,
         Long listaId,
         String titulo,
         String descricao,
